Timer mechanism update

+++ database/migrations/2016_01_14_003852_create_participant_exams_table.php
@@ -20,7 +20,7 @@ class CreateParticipantExamsTable extends Migration
             $table->integer('exam_id')
                 ->unsigned();
             $table->smallInteger('status');
-            $table->integer('score')
+            $table->decimal('score')
                 ->nullable();
             $table->integer('elapsed_time')
                 ->nullable();
