+++ database/migrations/2016_01_14_005326_create_participant_exams_answers_table.php
@@ -25,7 +25,7 @@ class CreateParticipantExamsAnswersTable extends Migration
             $table->text('answer_body')
                 ->nullable();
             $table->integer('elapsed_time');
-            $table->integer('score')
+            $table->decimal('score')
                 ->nullable();
             $table->timestamps();
             $table->softDeletes();
